Final submission with spelling in the test cases fixed
